Add trait mod
