+++ src/bin/02_plugin_drivers/domain/mod.rs
@@ -1,0 +1,1 @@
+pub mod traits;
